Fix: Moved create consultation mutation to Patient Symptoms so on error, the user is forwarded back the to enter code page.
