+++ client/src/Components/Pages/DoctorChat/DoctorChat.tsx
@@ -165,7 +165,7 @@ const DoctorChat = () => {
                   type="submit"
                   value="End consultation"
                   text="End consultation"
-                  onClick={() => {}}
+                  onClick={() => endConsultation()}
                 />
             </div>
           </div>
